pyproject.toml
[project]
name = "code_archive_wrapper"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "yzqzss", email = "yzqzss@yandex.com"},
]
dependencies = ["anyio>=4.9.0", "httpx>=0.28.1", "python-telegram-bot>=22.0", "dotenv>=0.9.9"]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = false
